+++ Package.swift
@@ -220,6 +220,7 @@ let package = Package(
                 "Models/HomeAttention.swift",
                 "Models/KeybindGrammar.swift",
                 "Models/KeybindScanner.swift",
+                "Models/KeybindOverview.swift",
                 "Models/MacKeyCodeMap.swift",
                 "Models/SavePlayerFields.swift",
                 "Models/SaveHeroPortrait.swift",